Add a roadmap and placeholders for auth checks
ShareJS has access to the request headers on connection (and any other
action).

Based on that we can use session cookies to carry authentication tokens.
The user logs in separately from ShareJS, the login process will set a
cookie which will then ShareJS test for authenticity and allow/reject
access based on that.