Fix creating new document breaking docs list
When we create a document and quickly go back to the list of all
documents, none of the docs will be shown.

The reason is that the `dbi.getDocument` function reads the snaphsot
from the database, but ShareJS doesn't update the snapshot immediatelly.

Therefore, instead of the new document we get `null`.

The proper solution would is to rewrite `dbi` to use ShareJS API rather
than reading the stuff directly from the database.