Don't access ShareJS database directly
It's brittle (the underlying storage can change any moment), inflexible
(if we want to switch from Redis to Postrges for ShareJS ops, we need to
rewrite it) and it doesn't work properly anyway (snapshots aren't
rebuild immediately so the database can have stale data).

We're now talking to the ShareJS model instead.