Fix the document-index hover issue
Fitts' law FTW.

In the document-index page, when the mouse moved between two items, none
were selected.

This makes it so that the whitespace is split and each half belongs to
one of the items. This effectively increases the active area of each
list item while keeping the same look.