Add development and production environments
Production serves ShareJS and API only and uses `settings.json` for
session secret and port number.

Dev uses default port and session secret settings and it serves the
Documara assets as well.

Production is expected to run in conjunction with a webserver (such as
Apache or nginx) that will serve the static assets.

Dev is standalone, but ineffective and insecure.

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>