Merge lists.html into documents.html
This will make the development quicker for now as we won't have to worry
about repeating the same markup for Documents and Lists (and any other
app) and we won't need any pre-processing tool.

We'll still keep the the javascript and css separate.

Later on, we can either do a proper single-page app and/or add some sort
of  asset processing tool.

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>