Add a key-based authentication to the API
The user is now abse to generate API keys (with optional description)
that can be used to authenticate against the API without using the
user's password.

Generating a new key:

    curl -X POST http://documara.com/api/keys/ -H 'Content-Type: application/json' \
    --user 'me@example.com:password' -d '{"name": "backup cron job"}'

This will return the key.

Listing all keys:

    curl http://documara.com/api/keys/ --user 'me@example.com:password'

Using the key:

    curl http://documara.com/api/documents/ \
    -H 'X-DOCUMARA-KEY: 53b9832828154268050f'

For security reasons, keys are not able to generate/list keys or work
with the login route.

Right now, the keys support read-only operations and don't have access
to ShareJS.

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>