Fix the session expiry
The session was expiring after a like a day or so, which was way to soon. If I
understand `connect.session` correctly, this should expire the session in 7 days
of inactivity. Meaning that if we use Documara at least once every 7 days, it
won't expire at all. Which is what I want.

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>