Increase Redis session TTL
This seems to be the piece we were missing for the early logout issues.
connect-redis stores the session keys with a TTL value of 1 day. Which
seems to be consistent with what we were seeing.

We've now set the value to 100 days. The cookies have a max expiry set
to 7 days and should refresh it on every access (though I suspect there
may be a bug there as well -- we'll see).